--- backend/src/main/java/com/powerme/dto/UpdatePasswordDto.java
@@ -1,24 +1,0 @@
-package com.powerme.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
-public class UpdatePasswordDto {
-
-    @NotBlank
-    @Size(min = 8)
-    private String newPassword;
-
-    public UpdatePasswordDto(String newPassword) {
-        this.newPassword = newPassword;
-    }
-
-    public String getNewPassword() {
-        return newPassword;
-    }
-
-    public void setNewPassword(String newPassword) {
-        this.newPassword = newPassword;
-    }
-}
-
